[Refactor] Improve module cohesion (Fixes #151)
