+++ tests/test_position_sizing.py
@@ -1,7 +1,7 @@
 import pytest
 
 from trading_bot.risk.position_sizing import calculate_position_size
-from trading_bot.risk_config import PositionSizingConfig
+from trading_bot.risk.config import PositionSizingConfig
 
 
 def test_fixed_cash_position_size():
@@ -30,7 +30,9 @@ def test_risk_per_trade_mode():
 
 def test_precision_flooring():
     cfg = PositionSizingConfig(mode="fixed_fraction", fraction_of_equity=0.5)
-    qty = calculate_position_size(cfg, price=3, equity=100, lot_size=0.1, precision=2)
+    qty = calculate_position_size(
+        cfg, price=3, equity=100, lot_size=0.1, precision=2
+    )
     assert qty == pytest.approx(16.6)
