+++ tests/test_risk_config.py
@@ -1,13 +1,10 @@
-import os
 import sys
 
 import pytest
-
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))
-from trading_bot.risk_config import get_risk_config
+from trading_bot.risk.config import get_risk_config
 from trading_bot.main import parse_args
-from trading_bot.utils.config import load_config
-from trading_bot.risk_config import (
+from trading_bot.config import load_config
+from trading_bot.risk.config import (
     PositionSizingConfig,
     StopLossConfig,
     TakeProfitConfig,
