+++ tests/test_validation.py
@@ -4,7 +4,7 @@ import sys
 import pytest
 
 from trading_bot.main import parse_args
-from trading_bot.utils.config import load_config
+from trading_bot.config import load_config
 
 
 def test_negative_trade_size_config(tmp_path):
