+++ trading_bot/risk/__init__.py
@@ -2,5 +2,12 @@
 
 from .position_sizing import calculate_position_size
 from .guardrails import Guardrails
+from .config import RiskConfig, PositionSizingConfig, get_risk_config
 
-__all__ = ["calculate_position_size", "Guardrails"]
+__all__ = [
+    "calculate_position_size",
+    "Guardrails",
+    "RiskConfig",
+    "PositionSizingConfig",
+    "get_risk_config",
+]
