+++ trading_bot/strategies/rsi_strategy.py
@@ -4,7 +4,7 @@ from typing import List, Dict, Any
 import numpy as np
 import pandas as pd
 
-from trading_bot.utils.config import get_config
+from trading_bot.config import get_config
 from trading_bot.strategies import register_strategy
 
 logger = logging.getLogger(__name__)
